Code Organization
+ WebsiteModel & SettingsItemProtocol refactored into their own files
+ SettingsItemProtocol Extension for configuring cells

- Removed SettingsItemModel
